+++ lib/cln.js
@@ -1,28 +1,9 @@
 import fetch from 'cross-fetch'
-import https from 'https'
 import crypto from 'crypto'
-import { HttpProxyAgent, HttpsProxyAgent } from '@/lib/proxy'
-import { TOR_REGEXP } from '@/lib/url'
+import { getAgent } from '@/lib/proxy'
 
 export const createInvoice = async ({ socket, rune, cert, label, description, msats, expiry }) => {
-  let protocol, agent
-  const httpsAgentOptions = { ca: cert ? Buffer.from(cert, 'base64') : undefined }
-  const isOnion = TOR_REGEXP.test(socket)
-  if (isOnion) {
-    // we support HTTP and HTTPS over Tor
-    protocol = cert ? 'https:' : 'http:'
-    // we need to use our Tor proxy to resolve onion addresses
-    const proxyOptions = { proxy: process.env.TOR_PROXY }
-    agent = protocol === 'https:'
-      ? new HttpsProxyAgent({ ...proxyOptions, ...httpsAgentOptions, rejectUnauthorized: false })
-      : new HttpProxyAgent(proxyOptions)
-  } else if (process.env.NODE_ENV === 'development' && !cert) {
-    protocol = 'http:'
-  } else {
-    // we only support HTTPS over clearnet
-    agent = new https.Agent(httpsAgentOptions)
-    protocol = 'https:'
-  }
+  const { protocol, agent } = getAgent({ hostname: socket, cert })
 
   const url = `${protocol}//${socket}/v1/invoice`
   const res = await fetch(url, {
