+++ lib/proxy.js
@@ -1,5 +1,6 @@
 import http from 'http'
 import https from 'https'
+import { TOR_REGEXP } from '@/lib/url'
 
 // from https://github.com/delvedor/hpagent
 
@@ -117,4 +118,29 @@ export class HttpsProxyAgent extends https.Agent {
 
     request.end()
   }
+}
+
+export function getAgent ({ hostname, cert }) {
+  let protocol, agent
+
+  const httpsAgentOptions = { ca: cert ? Buffer.from(cert, 'base64') : undefined }
+
+  const isOnion = TOR_REGEXP.test(hostname)
+  if (isOnion) {
+    // we support HTTP and HTTPS over Tor
+    protocol = cert ? 'https:' : 'http:'
+    // we need to use our Tor proxy to resolve onion addresses
+    const proxyOptions = { proxy: process.env.TOR_PROXY }
+    agent = protocol === 'https:'
+      ? new HttpsProxyAgent({ ...proxyOptions, ...httpsAgentOptions, rejectUnauthorized: false })
+      : new HttpProxyAgent(proxyOptions)
+  } else if (process.env.NODE_ENV === 'development' && !cert) {
+    protocol = 'http:'
+  } else {
+    // we only support HTTPS over clearnet
+    agent = new https.Agent(httpsAgentOptions)
+    protocol = 'https:'
+  }
+
+  return { protocol, agent }
 }
